Don't expose some classes
