+++ kafka-impl/src/main/java/io/streamnative/pulsar/handlers/kop/storage/ProducerStateManager.java
@@ -14,14 +14,123 @@
 package io.streamnative.pulsar.handlers.kop.storage;
 
 import com.google.common.collect.Maps;
+import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.TreeMap;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.experimental.Accessors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.common.record.RecordBatch;
 import org.apache.kafka.common.requests.FetchResponse;
+
+/**
+ * AbortedTxn is used cache the aborted index.
+ */
+@Data
+@Accessors(fluent = true)
+@AllArgsConstructor
+class AbortedTxn {
+
+    private static final int VersionOffset = 0;
+    private static final int VersionSize = 2;
+    private static final int ProducerIdOffset = VersionOffset + VersionSize;
+    private static final int ProducerIdSize = 8;
+    private static final int FirstOffsetOffset = ProducerIdOffset + ProducerIdSize;
+    private static final int FirstOffsetSize = 8;
+    private static final int LastOffsetOffset = FirstOffsetOffset + FirstOffsetSize;
+    private static final int LastOffsetSize = 8;
+    private static final int LastStableOffsetOffset = LastOffsetOffset + LastOffsetSize;
+    private static final int LastStableOffsetSize = 8;
+    private static final int TotalSize = LastStableOffsetOffset + LastStableOffsetSize;
+
+    private static final Short CurrentVersion = 0;
+
+    private final Long producerId;
+    private final Long firstOffset;
+    private final Long lastOffset;
+    private final Long lastStableOffset;
+
+    protected ByteBuffer toByteBuffer() {
+        ByteBuffer buffer = ByteBuffer.allocate(AbortedTxn.TotalSize);
+        buffer.putShort(CurrentVersion);
+        buffer.putLong(producerId);
+        buffer.putLong(firstOffset);
+        buffer.putLong(lastOffset);
+        buffer.putLong(lastStableOffset);
+        buffer.flip();
+        return buffer;
+    }
+}
+
+@Data
+@Accessors(fluent = true)
+@AllArgsConstructor
+class CompletedTxn {
+    private Long producerId;
+    private Long firstOffset;
+    private Long lastOffset;
+    private Boolean isAborted;
+}
+
+@Data
+@Accessors(fluent = true)
+@EqualsAndHashCode
+class TxnMetadata {
+    private final long producerId;
+    private final long firstOffset;
+    private long lastOffset;
+
+    public TxnMetadata(long producerId, long firstOffset) {
+        this.producerId = producerId;
+        this.firstOffset = firstOffset;
+    }
+}
+
+/**
+ * BatchMetadata is used to check the message duplicate.
+ */
+@Getter
+@AllArgsConstructor
+class BatchMetadata {
+
+    private final Integer lastSeq;
+    private final Long lastOffset;
+    // Should be seq delta, we use offsetDelta here because in future might change back.
+    // When we preset the correct offset before message publish.
+    private final Integer offsetDelta;
+    private final Long timestamp;
+
+    public int firstSeq() {
+        return decrementSequence(lastSeq, offsetDelta);
+    }
+
+    public Long firstOffset() {
+        return lastOffset - offsetDelta;
+    }
+
+    private int decrementSequence(int sequence, int decrement) {
+        if (sequence < decrement) {
+            return Integer.MAX_VALUE - (decrement - sequence) + 1;
+        }
+        return sequence - decrement;
+    }
+
+    @Override
+    public String toString() {
+        return "BatchMetadata("
+                + "firstSeq=" + firstSeq() + ", "
+                + "lastSeq=" + lastSeq + ", "
+                + "firstOffset=" + firstOffset() + ", "
+                + "lastOffset=" + lastOffset + ", "
+                + "timestamp=" + timestamp + ")";
+    }
+}
 
 /**
  * Producer state manager.
